+++ src/test/java/org/folio/marc/migrations/services/batch/saving/SavingRecordsChunkProcessorTest.java
@@ -5,12 +5,15 @@ import static org.folio.marc.migrations.domain.entities.types.EntityType.AUTHORI
 import static org.folio.marc.migrations.domain.entities.types.EntityType.INSTANCE;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.sql.Timestamp;
 import java.time.Instant;
+import java.util.List;
 import java.util.UUID;
+import java.util.stream.Stream;
 import org.assertj.core.api.SoftAssertions;
 import org.folio.marc.migrations.client.BulkClient.BulkResponse;
 import org.folio.marc.migrations.domain.entities.ChunkStep;
@@ -20,12 +23,16 @@ import org.folio.marc.migrations.domain.entities.types.OperationStatusType;
 import org.folio.marc.migrations.domain.entities.types.OperationStep;
 import org.folio.marc.migrations.domain.entities.types.StepStatus;
 import org.folio.marc.migrations.services.BulkStorageService;
+import org.folio.marc.migrations.services.batch.support.FolioS3Service;
 import org.folio.marc.migrations.services.domain.RecordsSavingData;
 import org.folio.marc.migrations.services.jdbc.ChunkStepJdbcService;
 import org.folio.spring.testing.type.UnitTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
@@ -39,6 +46,7 @@ class SavingRecordsChunkProcessorTest {
 
   private @Mock BulkStorageService bulkStorageService;
   private @Mock ChunkStepJdbcService chunkStepJdbcService;
+  private @Mock FolioS3Service s3Service;
   private @InjectMocks SavingRecordsChunkProcessor processor;
 
   @BeforeEach
@@ -70,15 +78,24 @@ class SavingRecordsChunkProcessorTest {
     save_positive(chunk, INSTANCE);
   }
 
-  @Test
-  void process_retry_shouldUpdateExistingChunkStep() {
+  @ParameterizedTest
+  @MethodSource("chunkArguments")
+  void process_retry_shouldUpdateExistingChunkStep(UUID operationId, EntityType entityType) {
     // Arrange
     int numOfRecords = 5;
-    var chunk = chunk(numOfRecords, AUTHORITY_OPERATION_ID, OperationStatusType.DATA_SAVING_FAILED);
+    var chunk = chunk(numOfRecords, operationId, OperationStatusType.DATA_SAVING_FAILED);
     var existingChunkStep = createChunkStep(chunk);
     when(chunkStepJdbcService.getChunkStepByChunkIdAndOperationStep(chunk.getId(), OperationStep.DATA_SAVING))
       .thenReturn(existingChunkStep);
-    processor.setEntityType(AUTHORITY);
+
+    var errorChunkFileName = existingChunkStep.getErrorChunkFileName();
+    var entityChunkFileName = chunk.getEntityChunkFileName();
+
+    // Mock S3 service behavior
+    when(s3Service.readFile(entityChunkFileName)).thenReturn(List.of("record1", "record2", "record3"));
+    when(s3Service.readFile(errorChunkFileName)).thenReturn(List.of("record1,error"));
+
+    processor.setEntityType(entityType);
     processor.setPublishEventsFlag(Boolean.TRUE);
 
     // Act
@@ -88,17 +105,21 @@ class SavingRecordsChunkProcessorTest {
     assertThat(result).isNotNull();
     verify(chunkStepJdbcService).updateChunkStep(eq(existingChunkStep.getId()), eq(StepStatus.IN_PROGRESS),
         any(Timestamp.class));
-    verify(bulkStorageService).saveEntities(chunk.getEntityChunkFileName(), AUTHORITY, Boolean.TRUE);
+    verify(s3Service).writeFile(entityChunkFileName, List.of("record1"));
+    verify(bulkStorageService).saveEntities(chunk.getEntityChunkFileName(), entityType, Boolean.TRUE);
   }
 
-  @Test
-  void process_retry_shouldCreateNewChunkStep_whenNoExistingChunkStep() {
+  @ParameterizedTest
+  @MethodSource("chunkArguments")
+  void process_retry_shouldCreateNewChunkStep_whenNoExistingChunkStep(UUID operationId, EntityType entityType) {
     // Arrange
     int numOfRecords = 5;
-    var chunk = chunk(numOfRecords, AUTHORITY_OPERATION_ID, OperationStatusType.DATA_SAVING_FAILED);
-    when(chunkStepJdbcService.getChunkStepByChunkIdAndOperationStep(chunk.getId(), OperationStep.DATA_SAVING))
-        .thenReturn(null);
-    processor.setEntityType(AUTHORITY);
+    var chunk = chunk(numOfRecords, operationId, OperationStatusType.DATA_SAVING_FAILED);
+    when(chunkStepJdbcService.getChunkStepByChunkIdAndOperationStep(chunk.getId(),
+        OperationStep.DATA_SAVING)).thenReturn(null);
+    doNothing().when(chunkStepJdbcService)
+      .createChunkStep(any());
+    processor.setEntityType(entityType);
     processor.setPublishEventsFlag(Boolean.TRUE);
 
     // Act
@@ -107,7 +128,7 @@ class SavingRecordsChunkProcessorTest {
     // Assert
     assertThat(result).isNotNull();
     verify(chunkStepJdbcService).createChunkStep(any());
-    verify(bulkStorageService).saveEntities(chunk.getEntityChunkFileName(), AUTHORITY, Boolean.TRUE);
+    verify(bulkStorageService).saveEntities(chunk.getEntityChunkFileName(), entityType, Boolean.TRUE);
   }
 
   void save_positive(OperationChunk chunk, EntityType entityType) {
@@ -172,9 +193,16 @@ class SavingRecordsChunkProcessorTest {
       .operationChunkId(chunk.getId())
       .operationStep(OperationStep.DATA_SAVING)
       .status(StepStatus.FAILED)
-      .numOfErrors(0)
-      .entityErrorChunkFileName(chunk.getEntityChunkFileName())
+      .numOfErrors(1)
+      .entityErrorChunkFileName("entity_error")
+      .errorChunkFileName("error")
       .stepStartTime(Timestamp.from(Instant.now()))
       .build();
+  }
+
+  private static Stream<Arguments> chunkArguments() {
+    return Stream.of(
+        Arguments.of(AUTHORITY_OPERATION_ID, AUTHORITY),
+        Arguments.of(INSTANCE_OPERATION_ID, INSTANCE));
   }
 }